src/backend/requirements.txt
fastapi[standard]
pymongo
python-dotenv
pytest
tensorflow
pandas
scikit-learn
xgboost